Implement improved logging functionality
Beginning to utilize the additional Severity and Source properties of
the LogMessage type. Using two dictionaries to make sense of the two
properties and update the display of output accordingly.

In addition, I also started using the Exception property to elaborate on
the cause of an error (if Message.Exception isn't null)