Add a new file for optional, extra commands
In this file I plan to put "neat" commands that don't interact with
Discord that much, such as the ability to send queries to web services
like Google and Wolfram Alpha.

I may also include a few extra bits that are a little bit of fun. The
plan is that server owners will be able to disable these commands and
have the bot purely for moderation/info purposes.

Google search will be achieved by using API keys, in which case we will
need a new token besides the one for Discord. Which brings me to my next
commit...