Attempt to reduce chat spam, moving errors to logs
Previously, the bot would reply with a message any time something went
wrong during the course of handling a command. This included unknown
command errors, which caused a great deal of spam. I was just following
along with the d.NET examples which had this, so I've decided to change
it to a warning in the console.

Later, I ought to evaluate what went wrong in the command and grade
severity appropriately, rather than just making it a blanket warning
(low severity as I figure most errors will be "unknown command").