Add dedicated Keychain class, extend logging functionality
For both changes, the idea is to make keys and logging more easily accessible outside of the Program class. As I understand it, the Log method is somewhat critical to discord.NET's operations, so I've merely extended it by adding my own version in there, which is static and can be used outside. I also moved the colors and prefixes there; as they are static the main Program class can draw from them still just fine, but they need to be in there to maintain full logging coverage for other classes.