Changed title URL to redirect to my CSE
It no longer redirects to the main Google page as I found that results are seldom reproducible between the two - likely because the main Google page is laden with algorithms analyzing your past usage and cookies, as well as
your location, and all that sort of stuff. Especially with the removal of the ability to specify more than three results (well, it wasn't really an ability as it never worked, but I digress) it is especially important to get reproducibility right.