Add voice channel dynamic renaming
The bot will rename the voice channel to the most 'popular' game within it, providing that this game has at least two people playing it. In cases where there is no single game with two common players, the channel's name is set to general. In cases where there are an equal number of 2 or more people playing the same game, the channel's name will be 'Multiple games'. The AFK channel will never be modified.

I understand this is really annoying for large servers, and is more suited to small servers like mine, so I will be adding settings in the forseeable future which will include a toggle for this behaviour, as well as other options.