Finalize settings frontend with message posting and addition of ratelimits
D.NET auto controls ratelimits, but I've been told it isn't flawless. To stay in Discord's good books, I've added a two second delay between any async action, be it posting a message or adding a reaction. This is twice what the normal ratelimit is. It's a bit slow, but this is a setup function so it's largely unimportant; it won't be run that frequently.