Make quicker work of reaction event handling (optimize)
Added a new list which tracks all setting message IDs, providing a far quicker way to know if a message whose reactions have been modified is a settings message, versus the LINQ query before. In implementing the new list, I also changed around the constructors for the higher-level settings types. The GuildBotSetting base constructor now does most of the work, so by calling the base constructor very little is done in the typed settings. That also fixes an issue where settings were being added to the list twice, probably because of a last-ditch debugging effort that was never removed.